% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{scale_colour_continuous_THF}
\alias{scale_colour_continuous_THF}
\title{THF default colours for colour option
Continuous colour scales THF for continuous variables}
\usage{
scale_colour_continuous_THF(...)
}
\description{
THF default colours for colour option
Continuous colour scales THF for continuous variables
}
\examples{

library(ggplot2)
ggplot(mtcars, aes(x=mpg, y=wt, colour=disp)) +
geom_point() +
scale_colour_continuous_THF() +
theme_THF()
}
