% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{scale_colour_THF}
\alias{scale_colour_THF}
\title{THF default colours
Discrete colour scales THF for categorical variables}
\usage{
scale_colour_THF()
}
\description{
THF default colours
Discrete colour scales THF for categorical variables
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(x=mpg, y=wt, colour=factor(cyl))) +
geom_point() +
scale_colour_discrete_THF() +
theme_THF()

}
