% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{scale_fill_continuous_THF}
\alias{scale_fill_continuous_THF}
\title{THF default colours for fill option
Continuous colour scales THF for continuous variables}
\usage{
scale_fill_continuous_THF(...)
}
\description{
THF default colours for fill option
Continuous colour scales THF for continuous variables
}
\examples{

library(ggplot2)
 ggplot(faithfuld, aes(waiting, eruptions)) + geom_raster(aes(fill = density)) +
 scale_fill_continuous_THF() + theme_THF()
}
