% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_THF}
\alias{theme_THF}
\title{Title}
\usage{
theme_THF(base_size = 11, base_family = "")
}
\arguments{
\item{base_size}{Font size}

\item{base_family}{Font}
}
\description{
Title
}
\examples{

library(ggplot2)

ggplot(mtcars, aes(x = mpg)) + geom_dotplot() + theme_THF()
}
