% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{THF_styles}
\alias{THF_styles}
\title{Title}
\usage{
THF_styles(base_size = 11, base_family = "LTUnivers")
}
\arguments{
\item{base_family}{}
}
\description{
Title
}
\examples{

library(ggplot2)

ggplot(mtcars, aes(x = mpg)) + geom_dotplot() + THF_styles()
}
