% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{scale_colour_continuous_THF}
\alias{scale_colour_continuous_THF}
\title{Continuous colour scales THF}
\usage{
scale_colour_continuous_THF()
}
\description{
Continuous colour scales THF
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(x=mpg, y=wt, colour=factor(cyl))) +
geom_point() +
scale_colour_discrete_THF() +
theme_THF()

}
