% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{scale_fill_THF}
\alias{scale_fill_THF}
\title{THF default colours for fill variable}
\usage{
scale_fill_THF()
}
\description{
THF default colours for fill variable
}
\examples{
library(ggplot2)

ggplot(ToothGrowth, aes(x=factor(dose), y=len, fill=factor(dose))) +
geom_boxplot() + scale_fill_THF() + theme_THF()

}
